requirements.txt
torch
opencv-python
numpy
pandas
facenet-pytorch
pillow
flask
firebase-admin
requests
gunicorn   # ✅ ADD THIS ONLY